Stub in bcReachabilityAnalyzer off ciMethod
The reachabilityAnalyzer walks the ciBytecodeStream to find the
classes that need to be loaded and the methods to process next.

It should have a similar design to the bcEscapeAnalyzer.

Add stubs to ciMethod to be able to create one.  The rest of the
code for this will be pulled across from the StaticAnalyzer in
the djh/exploration branch

Signed-off-by: Dan Heidinga <heidinga@redhat.com>